+++ jkorm/src/main/kotlin/com/tang/jkorm/annotation/Table.kt
@@ -8,4 +8,16 @@ package com.tang.jkorm.annotation
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.CLASS)
 @MustBeDocumented
-annotation class Table(val value: String)
+annotation class Table(
+
+    /**
+     * Table name
+     */
+    val value: String,
+
+    /**
+     * Table alias
+     */
+    val alias: String = ""
+
+)