+++ jkorm/src/main/kotlin/com/tang/jkorm/paginate/OrderItem.kt
@@ -1,7 +1,7 @@
 package com.tang.jkorm.paginate
 
 import com.tang.jkorm.function.SFunction
-import com.tang.jkorm.utils.Reflects.getColumnName
+import com.tang.jkorm.wrapper.Column
 import kotlin.reflect.KMutableProperty1
 
 /**
@@ -9,25 +9,18 @@ import kotlin.reflect.KMutableProperty1
  *
  * @author Tang
  */
-class OrderItem<T> {
+class OrderItem<T>(
 
-    var column: String = ""
+    var column: Column,
 
     var asc: Boolean = true
 
-    constructor(column: String, asc: Boolean = true) {
-        this.column = column
-        this.asc = asc
-    }
+) {
 
-    constructor(property: KMutableProperty1<T, *>, asc: Boolean = true) {
-        this.column = getColumnName(property)
-        this.asc = asc
-    }
+    constructor(column: String, asc: Boolean = true) : this(Column(column), asc)
 
-    constructor(function: SFunction<T, *>, asc: Boolean = true) {
-        this.column = getColumnName(function)
-        this.asc = asc
-    }
+    constructor(property: KMutableProperty1<T, *>, asc: Boolean = true) : this(Column(property), asc)
+
+    constructor(function: SFunction<T, *>, asc: Boolean = true) : this(Column(function), asc)
 
 }