+++ jkorm/src/main/kotlin/com/tang/jkorm/sql/provider/AbstractSqlProvider.kt
@@ -65,7 +65,8 @@ abstract class AbstractSqlProvider : SqlProvider {
         if (orderBys.isNotEmpty()) {
             sql.append(ORDER_BY)
             orderBys.joinToString(COMMA_SPACE) {
-                it.column + if (it.asc) ASC else DESC
+                // TODO check column name is correct
+                it.column.toString() + if (it.asc) ASC else DESC
             }.let { sql.append(it) }
         }
     }