+++ jkorm/src/main/kotlin/com/tang/jkorm/wrapper/query/JoinType.kt
@@ -1,0 +1,14 @@
+package com.tang.jkorm.wrapper.query
+
+/**
+ * @author Tang
+ */
+enum class JoinType {
+
+    LEFT,
+
+    RIGHT,
+
+    INNER
+
+}