+++ jkorm/src/test/kotlin/com/tang/jkorm/session/entity/Role.kt
@@ -1,11 +1,13 @@
 package com.tang.jkorm.session.entity
 
+import com.tang.jkorm.annotation.Table
 import com.tang.jkorm.annotation.id.Id
 import com.tang.jkorm.annotation.id.IdType
 
 /**
  * @author Tang
  */
+@Table(value = "role", alias = "re")
 class Role(
 
     @Id(type = IdType.GENERATOR)